Splitup erna.__init__
