+++ erna/__init__.py
@@ -1,221 +1,6 @@
-import logging
-import pandas as pd
-import os
-import numpy as np
-import json
 import pkg_resources
-from datetime import timedelta
-
-from . import datacheck_conditions as dcc
-from .datacheck import get_runs, get_drs_runs
-
-logger = logging.getLogger(__name__)
 
 
-def build_path(row, path_to_data, extension):
-    """
-    builds a path to the fact data given the night, extension and filename
-    """
-    night = str(row.night)
-    year = night[0:4]
-    month = night[4:6]
-    day = night[6:8]
-    res = os.path.join(path_to_data, year, month, day, row.filename + extension)
-    return res
-
-
-def test_drs_path(df, key):
-    """
-    Test if the given drs paths in the key are present
-    """
-    mask = df[key].apply(os.path.exists)
-    df['drs_file_exists'] = mask
-
-    return df
-
-
-def test_data_path(df, key):
-    """
-    Test the given data paths in key if they exists. It tests for
-    both possible fileextensions [.fz, .gz] and corrects if necessary.
-    """
-    mask = df[key].apply(os.path.exists)
-    df['data_file_exists'] = mask
-    df.loc[~mask, key] = df.loc[~mask, key].str.replace('.fz', '.gz')
-    df.loc[~mask, 'data_file_exists'] = df.loc[~mask, key].apply(os.path.exists)
-
-    return df
-
-
-def build_filename(night, run_id):
-    return night.astype(str) + '_' + run_id.map('{:03d}'.format)
-
-
-def mc_drs_file():
-    '''
-    return path to the drs file used for monte carlo files
-    '''
-    drs_path = pkg_resources.resource_filename(
-        __name__, 'resources/mc_drs_constants.drs.fits.gz'
-    )
-    return drs_path
-
-
-def ensure_output(output_path):
-    '''
-    Make sure the output file does not exist yet.
-    Create directorie to new output file if necessary
-    '''
-    if os.path.exists(output_path):
-        raise FileExistsError('The output file already exists.')
-    directory = os.path.dirname(output_path)
-    if directory:
-        os.makedirs(directory, exist_ok=True)
-
-
-
-
-def load(
-        earliest_night,
-        latest_night,
-        path_to_data,
-        factdb,
-        source_name="Crab",
-        timedelta_in_minutes=30,
-        data_conditions=dcc.conditions["standard"]
-        ):
-    '''
-    Given the earliest and latest night to fetch as a factnight string (as in 20141024)
-    this method returns a DataFrame containing the paths to data files
-    and their correpsonding .drs files.
-    The maximum time difference between the data and drs files is
-    specified by the timedelta_in_minutes parameter.
-
-    Returns None if no files can be found.
-    '''
-
-    logger.debug("Table names in DB: ")
-    logger.debug(factdb.table_names())
-
-    if len(factdb.table_names()) > 0:
-        logger.info("Connected to Database.")
-
-    logger.info("Reading Data from DataBase from {} to {} for source: {}".format(
-        earliest_night, latest_night, source_name
-    ))
-
-    conditions = [
-        'fNight >= {}'.format(earliest_night),
-        'fNight <= {}'.format(latest_night),
-        'fSourceName = "{}"'.format(source_name),
-    ]
-    conditions.extend(data_conditions)
-    logger.info('Querying data with conditions: {}'.format(' AND '.join(conditions)))
-    data = get_runs(
-        factdb,
-        conditions=conditions,
-        columns=(
-            'fNight AS night', 'fRunID AS run_id',
-            'fRunStart', 'fRunStop',
-            'fOnTime', 'fEffectiveOn',
-        ),
-    )
-
-    # now lets get all drs runs
-    drs_conditions = [
-        'fNight >= {}'.format(earliest_night),
-        'fNight <= {}'.format(latest_night),
-    ]
-
-    drs_data = get_drs_runs(
-        factdb, conditions=drs_conditions,
-        columns=('fNight AS night', 'fRunID AS run_id', 'fRunStart', 'fRunStop'),
-    )
-
-    if len(data) == 0 or len(drs_data) == 0:
-        logger.error('No data or drs files found that adhere to the specified query.')
-        return None
-
-    logger.info("Got {} data runs and {} runs".format(len(data), len(drs_data)))
-
-    # the timestamp should be unique for each observation.
-    # No two observations start at the same time
-    data.set_index("fRunStart", inplace=True)
-    drs_data.set_index("fRunStart", inplace=True)
-    # sorting data by their timestamp.
-    data = data.sort_index()
-    drs_data = drs_data.sort_index()
-
-    # write filenames
-    data["filename"] = build_filename(data.night, data.run_id)
-    drs_data["filename"] = build_filename(drs_data.night, drs_data.run_id)
-
-    # write path
-    data["path"] = data.apply(build_path, axis=1, path_to_data=path_to_data, extension='.fits.fz')
-    drs_data["path"] = drs_data.apply(build_path, axis=1, path_to_data=path_to_data, extension='.drs.fits.gz')
-
-    #remove all none existing drs files
-    drs_data = test_drs_path(drs_data, "path")
-    drs_data = drs_data[drs_data['drs_file_exists']]
-
-    # reindex the drs table using the index of the data table.
-    # There are always more data runs than drs run in the db.
-    # hence missing rows have to be filled either forward or backwards
-    earlier_drs_entries = drs_data.reindex(data.index, method="ffill")
-    earlier_drs_entries = earlier_drs_entries.fillna(axis="index", method="ffill")
-    later_drs_entries = drs_data.reindex(data.index, method="backfill")
-    later_drs_entries = later_drs_entries.fillna(axis="index", method="ffill")
-
-    # when backfilling the drs obeservations the last rows might be invalid and contain nans.
-    # We cannot drop them becasue the tables have to have the same length.
-    # in that case simply fill them up.
-    earlier_drs_entries["deltaT"] = np.abs(earlier_drs_entries.fRunStop - data.fRunStop)
-    later_drs_entries["deltaT"] = np.abs(later_drs_entries.fRunStop - data.fRunStop).fillna(axis='index', method='ffill')
-    d_later = later_drs_entries[later_drs_entries.deltaT < earlier_drs_entries.deltaT]
-    d_early = earlier_drs_entries[later_drs_entries.deltaT >= earlier_drs_entries.deltaT]
-
-    closest_drs_entries = pd.concat([d_early, d_later])
-    closest_drs_entries = closest_drs_entries[closest_drs_entries.deltaT < timedelta(minutes = timedelta_in_minutes)]
-
-    mapping = pd.concat([
-        closest_drs_entries.filename,
-        closest_drs_entries.path,
-        data.path,
-        closest_drs_entries.deltaT,
-        data.fOnTime, data.fEffectiveOn,
-        data.night,
-        data.run_id,
-    ], axis=1, keys=[
-        "filename",
-        "drs_path",
-        "data_path",
-        "delta_t",
-        "ontime",
-        "effective_on",
-        "night",
-        "run_id",
-    ])
-
-    mapping = mapping.dropna(how='any')
-
-    logger.info("Fetched {} data runs and approx {} drs entries from database where time delta is less than {} minutes".format(len(mapping), mapping['drs_path'].nunique(), timedelta_in_minutes))
-    # effective_ontime = (mapping['ontime'] * mapping['effective_on']).sum()
-    # logger.info("Effective on time: {}. Thats {} hours.".format(datetime.timedelta(seconds=effective_ontime), effective_ontime/3600))
-
-    return mapping
-
-
-def ft_json_to_df(json_path):
-    with open(json_path,'r') as text:
-        try:
-            logger.info("Reading fact-tools output.")
-            y = json.loads(text.read())
-            df_out = pd.DataFrame(y)
-            logger.info("Returning data frame with {} entries".format(len(df_out)))
-            return df_out
-        except ValueError:
-            logger.exception("Fact-tools output could not be read.")
-            return "error reading json"
-        except Exception:
-            logger.exception("Fact-tools output could not be gathered.")
-            return "error gathering output"
+mc_drs_file = pkg_resources.resource_filename(
+    __name__, 'resources/mc_drs_constants.drs.fits.gz'
+)
