+++ erna/automatic_processing/__init__.py
@@ -1,2 +1,9 @@
 from . import database
 from .database_utils import fill_data_runs, fill_drs_runs
+
+
+__all__ = [
+    'database',
+    'fill_drs_runs',
+    'fill_data_runs',
+]
