+++ erna/scripts/console.py
@@ -1,9 +1,6 @@
 from IPython import embed
 import click
 from datetime import date
-from tqdm import tqdm
-from threading import Thread
-from time import sleep
 
 from ..automatic_processing.database import *
 from ..automatic_processing.database_utils import *
